rename parameter parser to signature
I'm going to add more things that are specific to signatures.
